feat: refactor master data types and move to dedicated types file
